==> Notebooks/2. Data_Understanding_422982 (1).ipynb <==
import subprocess
import os
import pandas as pd
import requests
from bs4 import BeautifulSoup
import json
pd.set_option('display.max_rows', 500)

page=requests.get('https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/Fallzahlen.html')

soup=BeautifulSoup(page.content, 'html.parser')

html_table=soup.find('table')

all_rows=html_table.find_all('tr')

final_data_list=[]

for pos, rows in enumerate(all_rows):
    
    col_list=[each_col.get_text(strip=True) for each_col in rows.find_all('td')]
    final_data_list.append(col_list)

pd.DataFrame(final_data_list).dropna().rename(columns={0:'state'})

# section 3.4 Rest API Calls

data=requests.get('https://services7.arcgis.com/mOBPykOjAyBO2ZKk/arcgis/rest/services/Coronaf%C3%A4lle_in_den_Bundesl%C3%A4ndern/FeatureServer/0/query?outFields=*&where=1%3D1&f=geojson')

import json

json_objects=json.loads((data.content))

type(json_objects)

full_list=[]
for pos,each_dict in enumerate (json_objects['features'][:]):
    full_list.append(each_dict['properties'])

pd.DataFrame(full_list)
