==> Notebooks/3. Data_Preparation_422982 (1).ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime 
%matplotlib inline

import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.graph_objects as go
import plotly


mpl.rcParams['figure.figsize']=(16,9)
pd.set_option('display.max_rows', 500)
sns.set(style="darkgrid")

import plotly.graph_objects as go

datapath= 'https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv'
pd_raw=pd.read_csv('https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv')

pd_raw.head()

pd_raw.columns[:]

time_idx=pd_raw.columns[4:]

df_plot= pd.DataFrame({ 'Date':
                       time_idx
    
})
df_plot.head()

pd_raw['Country/Region'].unique()

pd_raw['Country/Region']

pd_raw[pd_raw['Country/Region']=='US']

import numpy as np
country_list=['Italy', 'US', 'Spain','Germany', 'Korea,South']

for each in country_list:
    df_plot[each]=np.array(pd_raw[pd_raw['Country/Region']==each].iloc[:,4::].sum(axis=0))

df_plot

df_plot.set_index('Date').plot()

# section 4.2 done Conversion of Date Objects

df_plot.head()

from datetime import datetime

df_plot.Date[0]

datetime.strptime(df_plot.Date[0], '%m/%d/%y')

time_idx=[datetime.strptime(each, '%m/%d/%y') for each in df_plot.Date]
time_str=[each.strftime('%Y-%m-%d') for each in time_idx]

time_idx

df_plot['Date']=time_idx
type(df_plot['Date'][0])

df_plot.head()

df_plot.to_csv('processed.csv', sep=';', index=False)

df_plot

# section 4.3 done Relational data structures

import pandas as pd

datapath= 'https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv'
pd_raw=pd.read_csv('https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv')

pd_raw.head()

pd_data_base=pd_raw.rename(columns={'Country/Region':'country','Province/State':'state' })

pd_data_base

pd_data_base=pd_data_base.drop(['Lat', 'Long'], axis=1)

pd_data_base.head()

pd_data_base.set_index(['state','country'])

test_pd=pd_data_base.set_index(['state','country']).T
test_pd

test_pd.columns

test_pd.stack(level=[0,1]).reset_index()

pd_relational_model=pd_data_base.set_index(['state', 'country'])\
.T \
.stack(level=[0,1]) \
.reset_index() \
.rename(columns={'level_0':'date', 0:'confirmed'},
       )

pd_relational_model.head()

pd_relational_model.dtypes

pd_relational_model.to_csv('COVID_relational_confirmed.csv', sep = ';', index = None)
