==> Notebooks/6. Data_Preparation_Large_File_422982 (1).ipynb <==
import pandas as pd
import numpy as np

from datetime import datetime

%matplotlib inline
pd.set_option('display.max_rows', 500)

pd_JH_data=pd.read_csv('COVID_relational_confirmed.csv',sep=';',parse_dates=[0])
pd_JH_data=pd_JH_data.sort_values('date',ascending=True).reset_index(drop=True).copy()
pd_JH_data.head()

test_data = pd_JH_data[((pd_JH_data['country'] == 'US')| 
                        (pd_JH_data['country'] == 'Germany')) & 
                       ((pd_JH_data['date'] > '2020-05-20') & (pd_JH_data['date'] < '2020-05-29'))]

test_data.head()
test_data.groupby(['country']).agg(np.max)

import numpy as np
import pandas as pd
from sklearn import linear_model
reg = linear_model.LinearRegression(fit_intercept = True)

def get_doubling_time_via_regression(in_array):
    ''' Use a linear regression to approximate the doubling rate'''
    y = np.array(in_array)
    X = np.arange(-1,2).reshape(-1,1)

    assert len(in_array) == 3
    reg.fit(X, y)
    intercept = reg.intercept_
    slope = reg.coef_

    return intercept/slope

def rolling_reg(df_input,col='confirmed'):
    ''' Input has to be a data frame'''
    ''' Return is single series (mandatory for group by apply)'''
    days_back=3
    result=df_input[col].rolling(
                window=days_back,
                min_periods=days_back).apply(get_doubling_time_via_regression,raw=False)
    return result

test_data[['state','country','confirmed']].groupby(['state','country']).apply(rolling_reg,'confirmed')

pd_DR_result=pd_JH_data[['state','country','confirmed']].groupby(['state','country']).apply(rolling_reg,'confirmed').reset_index()

pd_DR_result=pd_DR_result.rename(columns={'confirmed':'confirmed_DR', 'level_2':'index'})
pd_DR_result.head()

pd_JH_data=pd_JH_data.reset_index()
pd_JH_data.head()

pd_result_larg=pd.merge(pd_JH_data,pd_DR_result[['index','confirmed_DR']],on=['index'],how='left')
pd_result_larg.head()

from scipy import signal

def savgol_filter(df_input,column='confirmed',window=5):
    ''' Savgol Filter which can be used in groupby apply function 
        it ensures that the data structure is kept'''
    window=5, 
    degree=1
    df_result=df_input
    
    filter_in=df_input[column].fillna(0) # attention with the neutral element here
    
    result=signal.savgol_filter(np.array(filter_in),
                           5, # window size used for filtering
                           1)
    df_result[column+'_filtered']=result
    return df_result

pd_filtered_result = pd_JH_data[['state','country','confirmed']].groupby(['state','country']).apply(savgol_filter).reset_index()

pd_result_larg = pd.merge(pd_result_larg, pd_filtered_result[['index', 'confirmed_filtered']], on = ['index'], how = 'left')
pd_result_larg.head()

pd_filtered_doubling = pd_result_larg[['state','country','confirmed_filtered']].groupby(['state','country']).apply(rolling_reg, 'confirmed_filtered').reset_index()

pd_filtered_doubling = pd_filtered_doubling.rename(columns = {'level_2':'index', 'confirmed_filtered':'confirmed_filtered_DR'})

pd_filtered_doubling.tail()

pd_result_larg = pd.merge(pd_result_larg, pd_filtered_doubling[['index','confirmed_filtered_DR']], on = ['index'], how = 'left')
pd_result_larg.tail()

mask=pd_result_larg['confirmed']>100
pd_result_larg['confirmed_filtered_DR']=pd_result_larg['confirmed_filtered_DR'].where(mask, other=np.NaN)

pd_result_larg[pd_result_larg['country']=='Germany'].tail()

pd_result_larg.to_csv('COVID_final_set.csv',sep=';',index=False)
