==> Notebooks/7_Evaluation_Walk_Through_422982 (1).ipynb <==
# Check some parameters
# Depending where you launch your notebook, the relative path might not work
# You should start the notebook server from your base path
# When opening the notebook, typically your path will be ../ads_covid-19/notebooks
import os
if os.path.split(os.getcwd())[-1]=='notebooks':
    os.chdir("../")

'Your base path is at: '+os.path.split(os.getcwd())[-1]

# %load E:/ads_covid-19/IDS_covid-19/src/data/process_JH_data.py
import pandas as pd
import numpy as np

from datetime import datetime

def store_relational_JH_data():
    

    pd_raw = pd.read_csv('https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv')

    pd_data_base = pd_raw.rename(columns = {'Country/Region' : 'country', 'Province/State' : 'state'})

    pd_data_base['state'] = pd_data_base['state'].fillna('no')

    pd_data_base = pd_data_base.drop(['Lat', 'Long'], axis = 1)

    pd_relational_model = pd_data_base.set_index(['state', 'country']) \
                                    .T                                 \
                                    .stack(level = [0,1])              \
                                    .reset_index()                     \
                                    .rename(columns = {'level_0':'date', 0:'confirmed'})
    pd_relational_model['date'] = pd_relational_model['date'].astype('datetime64[ns]')
    pd_relational_model.to_csv('ads_covid-19_IDS_covid-19_data_processed_COVID_relational_confirmed.csv', sep = ';', index = None)
    print('Number of raws stored: '+str(pd_relational_model.shape[0]))

if __name__ == '__main__':
    store_relational_JH_data()

# %load E:/ads_covid-19/IDS_covid-19/src/features/build_features.py
import numpy as np
import pandas as pd
from sklearn import linear_model
reg = linear_model.LinearRegression(fit_intercept = True)
from scipy import signal

def get_doubling_time_via_regression(in_array):

    y = np.array(in_array)
    X = np.arange(-1,2).reshape(-1,1)

    assert len(in_array) == 3
    reg.fit(X, y)
    intercept = reg.intercept_
    slope = reg.coef_

    return intercept/slope

def savgol_filter(df_input,column='confirmed',window=5):
    ''' Savgol Filter which can be used in groupby apply function
        it ensures that the data structure is kept'''
    window=5,
    degree=1
    df_result=df_input

    filter_in=df_input[column].fillna(0) # attention with the neutral element here

    result=signal.savgol_filter(np.array(filter_in),
                           5, # window size used for filtering
                           1)
    df_result[column+'_filtered']=result
    return df_result

def rolling_reg(df_input,col='confirmed'):
    ''' input has to be a data frame'''
    ''' return is single series (mandatory for group by apply)'''
    days_back=3
    result=df_input[col].rolling(
                window=days_back,
                min_periods=days_back).apply(get_doubling_time_via_regression,raw=False)
    return result

def calc_filtered_data(df_input, filter_on = 'confirmed'):
    '''
        Calculates savgol filter and returns merged DataFrame
    '''
    must_contain = set(['state', 'country', filter_on])
    assert must_contain.issubset(set(df_input.columns))

    pd_filtered_result = df_input[['state','country',filter_on]].groupby(['state','country']).apply(savgol_filter).reset_index()
    df_output = pd.merge(df_input, pd_filtered_result[['index', filter_on +'_filtered']], on = ['index'], how = 'left')

    return df_output

def calc_doubling_rate(df_input, filter_on = 'confirmed'):
    '''
        Calculates approximated doubling rate and returns merged DataFrame
    '''
    must_contain = set(['state', 'country', filter_on])
    assert must_contain.issubset(set(df_input.columns))

    pd_DR_result = df_input[['state','country', filter_on]].groupby(['state','country']).apply(rolling_reg, filter_on).reset_index()
    pd_DR_result = pd_DR_result.rename(columns = {'level_2':'index', filter_on : filter_on+'_DR'})

    df_output = pd.merge(df_input, pd_DR_result[['index', filter_on+'_DR']], on = ['index'], how = 'left')
    return df_output


if __name__ == '__main__':
    test_data = np.array([2,4,6])
    result = get_doubling_time_via_regression(test_data)
    print('The test slope is: '+str(result))

    pd_JH_data=pd.read_csv('ads_covid-19_IDS_covid-19_data_processed_COVID_relational_confirmed.csv',sep=';',parse_dates=[0])
    pd_JH_data = pd_JH_data.sort_values('date',ascending=True).reset_index().copy()

    pd_result_large = calc_filtered_data(pd_JH_data)
    pd_result_large = calc_doubling_rate(pd_result_large)
    pd_result_large = calc_doubling_rate(pd_result_large, 'confirmed_filtered')
    print(pd_result_large.head())

# %load E:/ads_covid-19/IDS_covid-19/src/visualization/visualize.py
import numpy as np
import pandas as pd
import dash
dash.__version__
import dash_core_components as dcc
import dash_html_components as html
from dash.dependencies import Input, Output,State

import plotly.graph_objects as go

import os
print(os.getcwd())
df_input_large=pd.read_csv('COVID_final_set.csv',sep=';')


fig = go.Figure()

app = dash.Dash()
app.layout = html.Div([

    dcc.Markdown('''
    #  Applied Data Science on COVID-19 Data - Part 1

    * Goal of the project is to learn data science by applying a cross-industry standard process.\
        The default layout contains the confirmed infected cases in the log-scale format on the Y-axis\
        and Timeline in Days on the X-axis.

    ### The first dropdown menu enables selection of one or multiple  countries for visualization.
    ### The seconds dropdown menu contains four option:
        1. The ‘Timeline Confirmed’ represents confirmed infected cases along the timeline.
        2. The ‘Timeline Confirmed Filtered’ represents filtered(savgol filter) confirmed infected cases along the timeline.
        3. The ‘Timeline Doubling Rate’ represents the doubling rate of the infected cases along the timeline.
        4. The ‘Timeline Doubling Rate Filtered’ represents the doubling rate of the filtered(savgol filter) infected cases along the timeline.

    '''),

    dcc.Markdown('''
    ### Select country below:
    '''),


    dcc.Dropdown(
        id='country_drop_down',
        options=[ {'label': each,'value':each} for each in df_input_large['country'].unique()],
        value=['US', 'Germany','Italy'], # which are pre-selected
        multi=True
    ),

    dcc.Markdown('''
        ### Select timeline of confirmed COVID-19 cases or the approximated doubling time:
        '''),


    dcc.Dropdown(
    id='doubling_time',
    options=[
        {'label': 'Timeline Confirmed ', 'value': 'confirmed'},
        {'label': 'Timeline Confirmed Filtered', 'value': 'confirmed_filtered'},
        {'label': 'Timeline Doubling Rate', 'value': 'confirmed_DR'},
        {'label': 'Timeline Doubling Rate Filtered', 'value': 'confirmed_filtered_DR'},
    ],
    value='confirmed',
    multi=False
    ),

    dcc.Graph(figure=fig, id='main_window_slope')
])



@app.callback(
    Output('main_window_slope', 'figure'),
    [Input('country_drop_down', 'value'),
    Input('doubling_time', 'value')])
def update_figure(country_list,show_doubling):


    if 'DR' in show_doubling:
        my_yaxis={'type':"log",
               'title':'Approximated doubling rate over 3 days (larger numbers are better #stayathome)'
              }
    else:
        my_yaxis={'type':"log",
                  'title':'Confirmed infected people (source johns hopkins csse, log-scale)'
              }


    traces = []
    for each in country_list:

        df_plot=df_input_large[df_input_large['country']==each]

        if show_doubling=='doubling_rate_filtered':
            df_plot=df_plot[['state','country','confirmed','confirmed_filtered','confirmed_DR','confirmed_filtered_DR','date']].groupby(['country','date']).agg(np.mean).reset_index()
        else:
            df_plot=df_plot[['state','country','confirmed','confirmed_filtered','confirmed_DR','confirmed_filtered_DR','date']].groupby(['country','date']).agg(np.sum).reset_index()


        traces.append(dict(x=df_plot.date,
                                y=df_plot[show_doubling],
                                mode='markers+lines',
                                opacity=0.9,
                                name=each
                        )
                )

    return {
            'data': traces,
            'layout': dict (
                width=1280,
                height=720,

                xaxis={'title':'Timeline',
                        'tickangle':-45,
                        'nticks':20,
                        'tickfont':dict(size=14,color="#7f7f7f"),
                      },

                yaxis=my_yaxis
        )
    }

if __name__ == '__main__':

    app.run_server(debug=True, use_reloader=False)

# %load E:/ads_covid-19/IDS_covid-19/src/models/SIR.py
import pandas as pd
import numpy as np
from datetime import datetime
from scipy import optimize
from scipy import integrate

def SIR_model_t(SIR, t, beta, gamma):
    S,I,R=SIR
    dS_dt = -beta*I*S/N0
    dI_dt = beta*I*S/N0 - gamma*I
    dR_dt = gamma*I

    return dS_dt, dI_dt, dR_dt

def fit_odeint(x, beta, gamma):
    return integrate.odeint(SIR_model_t, (S0, I0, R0), x, args=(beta, gamma))[:,1]


if __name__ == '__main__':

    df_analyse=pd.read_csv('E:/ads_covid-19/IDS_covid-19/data/raw/COVID-19/csse_covid_19_data/SIR_raw.csv',sep=';')

    df_analyse.sort_values('Date',ascending=True)
    N0 = 1000000
    df_data = df_analyse[35:] ## We will consider data from 35th day, which is 26th Feb 2020
    t = np.arange(df_data.shape[0])

    for country in df_data.columns[1:]:
        ydata = np.array(df_data[df_data[country]>0][country])
        t = np.arange(len(ydata))
        I0=ydata[0]
        S0=N0-I0
        R0=0
        popt=[0.4,0.1]
        fit_odeint(t, *popt)
        popt, pcov = optimize.curve_fit(fit_odeint, t, ydata, maxfev=5000)
        perr = np.sqrt(np.diag(pcov))
        fitted=fit_odeint(t, *popt)
        f_padded = np.concatenate((np.zeros(df_data.shape[0]-len(fitted)) ,fitted)) #to make dimentions equal
        df_data[country + '_fitted'] = f_padded
    df_data.to_csv("E:/ads_covid-19/IDS_covid-19/data/processed/SIR_calculated.csv", sep = ';', index=False)
