==> Notebooks/11_World Map Plotly_dynamic.ipynb_422982 (1).ipynb <==
import numpy as np
import pandas as pd

import plotly.graph_objects as go
from plotly import tools

## check some parameters
## depending where you launch your notebook, the relative path might not work
## you should start the notebook server from your base path
## when opening the notebook, typically your path will be ../ads_covid-19/notebooks
import os
if os.path.split(os.getcwd())[-1]=='notebooks':
    os.chdir("../")

'Your base path is at: '+os.path.split(os.getcwd())[-1]

url="https://covid.ourworldindata.org/data/owid-covid-data.csv"
df_country_info=pd.read_csv(url,sep=',')
#df_country_info=pd.read_csv('data/external/20220102 owid-covid-data.csv',sep=',')
df_country_info.head(9)

df_country_info.columns

# cut out the last day with data
df_country_info_day=df_country_info[df_country_info['date']==max(df_country_info['date'])]
df_country_info_day['total_cases_per_million']

#visualize_col='total_vaccinations_per_hundred'
visualize_col='total_boosters_per_hundred'
#df_country_info[['iso_code','date',visualize_col]].fillna(0)
#my_z=df_country_info[['iso_code',visualize_col]].groupby(['iso_code']).max().reset_index()
#my_z=df_country_info_day['total_cases_per_million']/10000
my_z=df_country_info_day[visualize_col]

fig = go.Figure(go.Choropleth(
                    locations = df_country_info_day['iso_code'],
                    z = my_z, # norm to percent
                    text = df_country_info_day['location'],
                    colorscale = 'Blues',
                    autocolorscale=False,
                    reversescale=False,
                    marker_line_color='darkgray',
                    marker_line_width=0.5,
                    colorbar_title = visualize_col,
                    ))
fig.show()

fig.update_layout(
    width=900, height=600,
    title_text='COVID-19 Total cases',
    geo=dict(
        showframe=False,
        showcoastlines=False,
        projection_type='equirectangular'
    ),
    annotations = [dict(
        x=0.55,
        y=0.1,
        xref='paper',
        yref='paper',
        text='Source: <a href="https://github.com/owid/covid-19-data/tree/master/public/data/">\
            Our World in Data</a>',
        showarrow = False
    )]
)

fig.show()
fig.write_html("Covid_world_map.html")

import plotly.graph_objects as go
from plotly import tools
#visualize_col='total_vaccinations_per_hundred'
visualize_col='total_boosters_per_hundred'
all_dates=df_country_info['date'].unique()
all_dates.sort()
my_z=df_country_info[['iso_code','date',visualize_col]].groupby(['iso_code']).max().reset_index()
#all_dates=all_dates[all_dates>'2021-01-01']
all_dates
